test format and display in cli/tests/unit/jupyter_test.ts
